Refactor user editing functionality by replacing UserEditPod with UserScene and introducing tabbed navigation for user details and password reset.
