+++ src/modules/users/edit-reset-password/edit-reset-password.pod.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const EditResetPassword: React.FC = () => {
+  return <h3>Reset Password</h3>;
+};
