+++ src/modules/users/edit-reset-password/index.ts
@@ -1,0 +1,1 @@
+export * from './edit-reset-password.pod';
