--- src/modules/users/edit/index.ts
@@ -1,1 +1,0 @@
-export * from './user-edit.pod';
