--- src/modules/users/edit/user-edit.styles.ts
@@ -1,8 +1,0 @@
-import { css } from '@emotion/css';
-
-export const root = css`
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  align-items: center;
-`;
