+++ src/scenes/_auth/edit-user/components/tab-panel.styles.ts
@@ -1,0 +1,6 @@
+import { css } from '@emotion/css';
+import { Theme } from '@mui/material';
+
+export const tabPanel = (theme: Theme) => css`
+  padding-block: ${theme.spacing(2)};
+`;
